venom
"To be honest, I didn't achieve what I initially aimed for, and I'm quite far from it. However, in an attempt to create something creative and fun, I thought it might be interesting to develop an online, network-based 'pet' that can visually react to sound waves (music). Imagine being able to see the frequencies of the music you're listening to, visualized as Venom (from the famous movie starring Tom Hardy) in real-time. This could offer a deeper understanding of music, allowing you to 'see' the sound, not just hear it!"